Convert to new hook system
