+++ includes/SkinFemiwiki.php
@@ -287,13 +287,13 @@ class SkinFemiwiki extends SkinMustache {
 	public function initPage( OutputPage $out ) {
 		$out->addMeta( 'viewport', 'width=device-width, initial-scale=1.0' );
 
-		$twitter = $this->getConfig()->get( 'FemiwikiTwitterAccount' );
+		$twitter = $this->getConfig()->get( Constants::CONFIG_TWITTER_ACCOUNT );
 		if ( $twitter ) {
 			$out->addMeta( 'twitter:site', "@$twitter" );
 		}
 
 		// Favicons
-		$headItems = $this->getConfig()->get( 'FemiwikiHeadItems' );
+		$headItems = $this->getConfig()->get( Constants::CONFIG_HEAD_ITEMS );
 		if ( $headItems ) {
 			$out->addHeadItems( $headItems );
 		}
@@ -305,7 +305,7 @@ class SkinFemiwiki extends SkinMustache {
 
 	/** @return string|null */
 	private function getAddThisPubId() {
-		$config = $this->getConfig()->get( 'FemiwikiAddThisId' );
+		$config = $this->getConfig()->get( Constants::CONFIG_ADD_THIS_ID );
 		if ( !$config ) {
 			return null;
 		}
